requirements-in.txt
numpy
scipy
pillow
opencv-python>=4.8.0
imageio
imageio-ffmpeg
adbutils
uiautomator2
uiautomator2cache
wrapt
retrying
jellyfish
pyyaml
inflection
pywebio
starlette==0.49.1
anyio
uvicorn[standard]
aiofiles
pypresence
rich
zerorpc
pyzmq
onepush
pydantic
psutil
chardet
matplotlib
pycryptodome
typing_extensions>=4.8.0
pydantic>=2.5.0
rapidocr
onnxruntime-directml>=1.24.4; sys_platform == "win32"
onnxruntime>=1.24.4; sys_platform == "linux"
onnxruntime @ https://files.pythonhosted.org/packages/fb/aa/04530bd38e31e26970fa1212346d76cf81705dc16a8ee5e6f4fb24634c11/onnxruntime-1.25.1-cp314-cp314-macosx_14_0_arm64.whl; sys_platform == "darwin" and platform_machine == "arm64"
watchdog>=2.0.0
numba
pip
lz4
setuptools==75.8.2
importlib_resources>=6.0.0
importlib_metadata>=8.0.0
openai
mcp==1.23.0
sse-starlette==3.0.3
pygments>=2.20.0
